Split web/main.py into five focused modules
web/main.py shrinks from 1,114 → 295 LOC (app construction + 13 route
handlers + _filter_response). All other logic moves to:

- web/filters.py (76 LOC): FilterSpec, MarkerFilter, VALID_TABS,
  _resolve_tab. Pure data; no web/state imports.
- web/state.py (207 LOC): AppState, state singleton, _load_demo,
  filtered-data LRU cache, age/marker range helpers, _normalise_age,
  _display_unit, MULTI_UNIT_MARKERS, _filter_ui_context,
  _units_ui_context. Eagerly loads the demo cohort on import.
- web/charts.py (262 LOC): CLUSTER_COLOURS, CHART_FONT, _base_layout,
  _embed, and the four Plotly chart builders.
- web/contexts.py (349 LOC): the six _*_context functions, _common,
  _build_tab_ctx. Wires chart HTML into Jinja-ready dicts.

web/__init__.py now adds the project root to sys.path so the new
submodules can import analysis/parsing/thresholds/unit_conversions
without each repeating the hack.

Dependency graph (no cycles):
  filters → ∅
  state   → filters
  charts  → state
  contexts → state, filters, charts
  main    → state, filters, charts, contexts

Two functions moved off the issue's suggested layout: _filter_ui_context
and _units_ui_context are state-driven (read state.df_long_full and
state.unit_prefs) so they sit in state.py alongside _display_unit and
MULTI_UNIT_MARKERS rather than contexts.py. This keeps every module
under the 350-LOC acceptance criterion.

tests/test_web.py imports updated to point at canonical new locations
(web.state for state/_load_demo/_filtered_data_cached/MULTI_UNIT_MARKERS,
web.main for the FastAPI app). All 228 tests still pass; manual smoke
test confirms every route returns 200.

Closes #21
